terminal_mock (cellium v0.1.0)

A mock terminal implementation that stores the screen state in a Map buffer. This allows for verifying what was drawn to the screen during tests.

Functions

code_change(OldVsn, State, Extra)

dump_screen()

get_cell(X, Y)

get_screen()

handle_call/3

handle_cast(Msg, State)

handle_info(Info, State)

init/1

start_link()

stop()

term_clear()

term_get_row(Y, Width)

term_get_row(X, Y, Length)

term_height()

term_init()

term_poll_event()

term_present()

term_print(X, Y, Fg, Bg, Str)

term_set_cell(X, Y, Char, Fg, Bg)

term_set_input_mode/1

term_set_output_mode/1

term_shutdown()

term_width()

terminate(Reason, State)